Refactor DefaultCodeActionProvider
